# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SwipeFash - Fashion E-Commerce Swipe API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Product {
  id: number;
  name: string;
  brand: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  description: string;
  imageUrl: string;
  images: string[];
  category: string;
  sizes: string[];
  colors: string[];
  stock: number;
  /** @nullable */
  rating?: number | null;
  reviewCount: number;
  isNew: boolean;
  isSale: boolean;
  createdAt: string;
}

export interface ProductFeed {
  products: Product[];
  /** @nullable */
  nextCursor?: number | null;
  total: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface ActivityItem {
  orderId: number;
  productName: string;
  action: string;
  timestamp: string;
}

export interface ProductStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  categoryBreakdown: CategoryCount[];
  recentActivity: ActivityItem[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  productCount: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  unpaid: 'unpaid',
  paid: 'paid',
  refunded: 'refunded',
} as const;

export interface Order {
  id: number;
  sessionId: string;
  productId: number;
  product: Product;
  selectedSize: string;
  selectedColor: string;
  quantity: number;
  totalPrice: number;
  status: OrderStatus;
  /** @nullable */
  paymentMethod?: string | null;
  paymentStatus: OrderPaymentStatus;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderInput {
  sessionId: string;
  productId: number;
  selectedSize: string;
  selectedColor: string;
  quantity: number;
}

export interface OrderConfirm {
  paymentMethod: string;
  shippingAddress: string;
  customerName: string;
  customerEmail: string;
}

export type ListProductsParams = {
category?: string;
cursor?: number;
limit?: number;
};

export type ListOrdersParams = {
session_id?: string;
};