# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SwipeFash - Fashion E-Commerce Swipe API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Fashion product operations
  - name: orders
    description: Order management
  - name: categories
    description: Product categories
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products for swipe feed
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductFeed"

  /products/stats:
    get:
      operationId: getProductStats
      tags: [products]
      summary: Get product statistics
      responses:
        "200":
          description: Product stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductStats"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders by session
      parameters:
        - name: session_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order after swipe right
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Out of stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/confirm:
    patch:
      operationId: confirmOrder
      tags: [orders]
      summary: Confirm payment and finalize order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderConfirm"
      responses:
        "200":
          description: Order confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/cancel:
    patch:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        description:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        isNew:
          type: boolean
        isSale:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - brand
        - price
        - description
        - imageUrl
        - images
        - category
        - sizes
        - colors
        - stock
        - reviewCount
        - isNew
        - isSale
        - createdAt

    ProductFeed:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        nextCursor:
          type: ["integer", "null"]
        total:
          type: integer
      required:
        - products
        - total

    ProductStats:
      type: object
      properties:
        totalProducts:
          type: integer
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        categoryBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required:
        - totalProducts
        - totalOrders
        - totalRevenue
        - categoryBreakdown
        - recentActivity

    CategoryCount:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required:
        - category
        - count

    ActivityItem:
      type: object
      properties:
        orderId:
          type: integer
        productName:
          type: string
        action:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - orderId
        - productName
        - action
        - timestamp

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
        - productCount

    Order:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        selectedSize:
          type: string
        selectedColor:
          type: string
        quantity:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        paymentMethod:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [unpaid, paid, refunded]
        shippingAddress:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - productId
        - product
        - selectedSize
        - selectedColor
        - quantity
        - totalPrice
        - status
        - paymentStatus
        - createdAt
        - updatedAt

    OrderInput:
      type: object
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        selectedSize:
          type: string
        selectedColor:
          type: string
        quantity:
          type: integer
      required:
        - sessionId
        - productId
        - selectedSize
        - selectedColor
        - quantity

    OrderConfirm:
      type: object
      properties:
        paymentMethod:
          type: string
        shippingAddress:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
      required:
        - paymentMethod
        - shippingAddress
        - customerName
        - customerEmail