# lib/api-zod/src/generated/types/categoryCount.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SwipeFash - Fashion E-Commerce Swipe API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryCount {
  category: string;
  count: number;
}