# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SwipeFash - Fashion E-Commerce Swipe API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';
import type { Product } from './product';

export interface Order {
  id: number;
  sessionId: string;
  productId: number;
  product: Product;
  selectedSize: string;
  selectedColor: string;
  quantity: number;
  totalPrice: number;
  status: OrderStatus;
  /** @nullable */
  paymentMethod?: string | null;
  paymentStatus: OrderPaymentStatus;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  createdAt: Date;
  updatedAt: Date;
}