# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SwipeFash - Fashion E-Commerce Swipe API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  brand: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  description: string;
  imageUrl: string;
  images: string[];
  category: string;
  sizes: string[];
  colors: string[];
  stock: number;
  /** @nullable */
  rating?: number | null;
  reviewCount: number;
  isNew: boolean;
  isSale: boolean;
  createdAt: Date;
}