# lib/api-zod/src/generated/types/productFeed.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SwipeFash - Fashion E-Commerce Swipe API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface ProductFeed {
  products: Product[];
  /** @nullable */
  nextCursor?: number | null;
  total: number;
}