# lib/api-zod/src/generated/types/productStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SwipeFash - Fashion E-Commerce Swipe API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';
import type { CategoryCount } from './categoryCount';

export interface ProductStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  categoryBreakdown: CategoryCount[];
  recentActivity: ActivityItem[];
}